nxp-imx9: remove real time related patches
Issue: LINUXEXEC-33080

Currently, in wrlinux1024 that is based on yocto 5.0. the ethtool
version is v6.7, and linuxptp version is v4.1. These real time
related patches have not been applied directly, so remove these
patches temporarily. Moreover, it only needs to basic imx93 BSP
in RCPL2, so the real time feature will be added in later when NXP
releases the real time SDK based on yocto 5.0.

Signed-off-by: Meng Li <Meng.Li@windriver.com>